requirements.txt
Flask
python-dotenv
gunicorn
flask-dance
werkzeug